ContextAPIを使ったグローバルステート化
