+++ firebase-react-hosting/src/components/TodoList.js
@@ -1,7 +1,9 @@
-import React from 'react';
+import React, { useContext } from 'react';
+import { TodosContext } from '../context/TodosContext';
 import { Button, Table } from 'reactstrap';
 
-function TodoList({ todos, completeTodo, removeTodo}) {
+function TodoList() {
+  const {todos,completeTodo,removeTodo}=useContext(TodosContext)
   return (
     <Table>
       <tbody>
